backend/requirements.txt
# MySindic - Dépendances Python
# Date: 24 octobre 2025

# Framework Web
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0

# Base de données
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentification et Sécurité
PyJWT==2.8.0
Werkzeug==3.0.1
cryptography==41.0.7

# Email
Flask-Mail==0.9.1

# Utilitaires
python-dotenv==1.0.0
gunicorn==21.2.0

# Validation de données
email-validator==2.1.0
